pyproject.toml
[tool.poetry]
name = "toco-discord-bot"
version = "0.1.0"
description = "My personal Discord bot"
authors = ["Stefano Di Pasquale <stefaniscion@gmail.com>"]
readme = "README.md"
packages = [{ include = "toco" }]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"

[tool.poetry.scripts]
toco = "toco.main:main"